+++ tensor2tensor/rl/collect_benchmark.py
@@ -23,7 +23,7 @@ from tensorflow.contrib import tpu
 from tensorflow.contrib.cluster_resolver import TPUClusterResolver
 
 from tensor2tensor.data_generators.gym_env import DummyWorldModelProblem
-from tensor2tensor.rl import ppo
+from tensor2tensor.rl import ppo_tpu
 from tensor2tensor.rl import tf_new_collect
 from tensor2tensor.utils import registry
 from tensor2tensor.utils import trainer_lib
@@ -47,46 +47,6 @@ tf.flags.DEFINE_string(
 HISTORY = 4
 
 
-# This runs just the video prediction model.
-#def make_model_fn(batch_size, epoch_length, num_tpus):
-#  def model_fn(features, labels, mode, params):
-#    with tf.variable_scope("model", reuse=tf.AUTO_REUSE):
-#      model_hparams = trainer_lib.create_hparams("next_frame_basic_stochastic_discrete_long")
-#      problem = DummyWorldModelProblem(
-#        action_space=Discrete(2), reward_range=(-1, 1),
-#        frame_height=210, frame_width=160
-#      )
-#      trainer_lib.add_problem_hparams(model_hparams, problem)
-#      model_hparams.force_full_predict = True
-#      model = registry.model("next_frame_basic_stochastic_discrete")(
-#          model_hparams, tf.estimator.ModeKeys.PREDICT
-#      )
-#      
-#      stack_size = model_hparams.video_num_input_frames
-#      history = tf.zeros(
-#          (batch_size, stack_size, problem.frame_height, problem.frame_width, 3),
-#          dtype=tf.int64
-#      )
-#      action = tf.zeros((batch_size, stack_size), dtype=tf.int64)
-#      
-#      with tf.variable_scope(tf.get_variable_scope(), reuse=tf.AUTO_REUSE):
-#        model.hparams.video_num_target_frames = 1
-#        model_output = model.infer({
-#            "inputs": history,
-#            "input_action": action,
-#            "reset_internal_states": 0.0
-#        })
-#
-#    # Summaries don't work when training on TPU; remove them.
-#    tf.get_default_graph().get_collection_ref(tf.GraphKeys.SUMMARIES)[:] = []
-#    return tf.contrib.tpu.TPUEstimatorSpec(
-#        mode=tf.estimator.ModeKeys.PREDICT,
-#        predictions={"x": model_output["targets"]},
-#    )
-#  return model_fn
-
-
-# This runs the new collect.
 def make_model_fn(epoch_length, num_tpus):
   def model_fn(features, labels, mode, params):
     batch_size = features.shape[0]
@@ -109,7 +69,8 @@ def make_model_fn(epoch_length, num_tpus):
         memory = tf_new_collect.new_define_collect(
             batch_env, ppo_hparams, action_space
         )
-      ppo_hparams.optimization_batch_size = 32
+      # Overriding the batch size to the minimal value to speed this up.
+      ppo_hparams.optimization_batch_size = 1
 
       ppo_summary = ppo.define_ppo_epoch(
           memory, ppo_hparams, action_space, batch_env.tensor_specs.observation
@@ -118,15 +79,10 @@ def make_model_fn(epoch_length, num_tpus):
     # Summaries don't work when training on TPU; remove them.
     tf.get_default_graph().get_collection_ref(tf.GraphKeys.SUMMARIES)[:] = []
 
-    #train_op = memory.reward
-    #train_op = tf.reduce_sum(ppo_summary)
-
     return tf.contrib.tpu.TPUEstimatorSpec(
-        #mode=tf.estimator.ModeKeys.TRAIN,
-        #loss=0.0,
-        #train_op=train_op,
         mode=tf.estimator.ModeKeys.PREDICT,
         predictions={"x": tf.expand_dims(ppo_summary[0], axis=0)},
+        # Switch to this variant to run just the collect, without PPO.
         #predictions={"x": tf.math.reduce_sum(memory[1], axis=-1)},
     )
   return model_fn
